pyproject.toml
[project]
name = "medical-rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite==0.19.0",
    "alembic==1.13.1",
    "asyncpg>=0.30.0",
    "fastapi==0.104.1",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "langchain==0.0.340",
    "langchain-openai==0.0.2",
    "numpy>=1.26.0,<2.0.0",
    "openai>=1.6.1,<2.0.0",
    "pandas==2.1.3",
    "pgvector>=0.4.1",
    "pinecone-client==2.2.4",
    "psycopg2-binary>=2.9.10",
    "pydantic==2.5.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "scikit-learn==1.3.2",
    "sqlalchemy==2.0.23",
    "tenacity>=8.2.0",
    "tiktoken>=0.5.1",
    "tqdm==4.66.5",
    "uvicorn[standard]==0.24.0",
]
